trading-bot/requirements.txt
ccxt==4.4.54
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pandas_ta==0.3.14b0
pyyaml==6.0.2
requests==2.32.3
